Updater work
Removed check for latest version at rconfig.com to github.  Removed
requirement for latest version to be the one being uploaded.  Added
option if server isn't connected to the internet or if the uploaded file
isn't the latest version to bypass this check and force rConfig to
update via the uploaded file.  Not fully finished yet.